+++ src/pages/home/home.page.spec.ts
@@ -1,32 +1,50 @@
 import { HomePage } from './home.page';
-import { TestBed, inject } from '@angular/core/testing';
-import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
-import { mockNavController } from 'ionic-angular/util/mock-providers';
+import { TestBed, ComponentFixture } from '@angular/core/testing';
+import { CUSTOM_ELEMENTS_SCHEMA, DebugElement } from '@angular/core';
 import { NavController } from 'ionic-angular';
 
 describe('Home Page:', () => {
 
-    beforeEach(() => TestBed.configureTestingModule({
-        schemas: [CUSTOM_ELEMENTS_SCHEMA],
-        providers: [
-            {provide: NavController, useValue: mockNavController},
-            HomePage,
-        ],
-    }));
+    let comp: HomePage;
+    let fixture: ComponentFixture<HomePage>;
+    let de: DebugElement;
 
-    it('Barney should be the user', inject([HomePage], (home: HomePage) => {
-        expect(home.user).toEqual({
+    beforeEach(() => {
+        TestBed.configureTestingModule({
+            schemas: [CUSTOM_ELEMENTS_SCHEMA],
+            declarations: [HomePage],
+            providers: [
+                {provide: NavController, useValue: NavController},
+            ],
+        });
+        fixture = TestBed.createComponent(HomePage);
+        // #trick
+        // If you want to trigger ionViewWillEnter automatically de-comment the following line
+        // fixture.componentInstance.ionViewWillEnter();
+        fixture.detectChanges();
+        comp = fixture.componentInstance;
+        de = fixture.debugElement;
+    });
+
+    describe('.constructor()', () => {
+        it('Should be defined', () => {
+            expect(comp).toBeDefined();
+        });
+    });
+
+    it('Barney should be the user', () => {
+        expect(comp.user).toEqual({
             name: 'barney',
             age: 36,
             active: true,
         });
-    }));
+    });
 
-    it('Fred should not be the user', inject([HomePage], (home: HomePage) => {
-        expect(home.user).not.toEqual({
+    it('Fred should not be the user', () => {
+        expect(comp.user).not.toEqual({
             name: 'fred',
             age: 36,
             active: true,
         });
-    }));
+    });
 });