+++ CHANGELOG.md
@@ -2,6 +2,85 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="0.18.0"></a>
+# [0.18.0](https://github.com/marcoturi/ionic2-boilerplate/compare/v0.17.0...v0.18.0) (2017-01-11)
+
+
+### Bug Fixes
+
+* **test:** add options in karma.conf and webpack.test to enanche the console log outputs during unit ([9b6c64a](https://github.com/marcoturi/ionic2-boilerplate/commit/9b6c64a))
+
+
+### Features
+
+* **git:** remove ghooks in favor of husky ([cf4337d](https://github.com/marcoturi/ionic2-boilerplate/commit/cf4337d))
+* **npm:** update ionic and other dips ([5bfddec](https://github.com/marcoturi/ionic2-boilerplate/commit/5bfddec))
+
+
+
+<a name="0.17.0"></a>
+# [0.17.0](https://github.com/marcoturi/ionic2-boilerplate/compare/v0.16.0...v0.17.0) (2016-12-20)
+
+
+### Bug Fixes
+
+* **npm:** stick karma-remap-coverage to 0.12, close [#20](https://github.com/marcoturi/ionic2-boilerplate/issues/20) ([2251eec](https://github.com/marcoturi/ionic2-boilerplate/commit/2251eec))
+
+
+### Features
+
+* **npm:** update to ionic app scripts ([7e26c68](https://github.com/marcoturi/ionic2-boilerplate/commit/7e26c68))
+
+
+
+<a name="0.16.0"></a>
+# [0.16.0](https://github.com/marcoturi/ionic2-boilerplate/compare/v0.15.1...v0.16.0) (2016-12-16)
+
+
+### Bug Fixes
+
+* **tests:** add angular2-template-loader to load templateUrls [#17](https://github.com/marcoturi/ionic2-boilerplate/issues/17) ([6fab1e1](https://github.com/marcoturi/ionic2-boilerplate/commit/6fab1e1))
+
+
+### Features
+
+* **npm:** update to ionic rc4 and tslint 4 ([d3410f7](https://github.com/marcoturi/ionic2-boilerplate/commit/d3410f7))
+
+
+### Reverts
+
+* **npm:** revert app-script version ([626bf5a](https://github.com/marcoturi/ionic2-boilerplate/commit/626bf5a))
+
+
+
+<a name="0.15.1"></a>
+## [0.15.1](https://github.com/marcoturi/ionic2-boilerplate/compare/v0.15.0...v0.15.1) (2016-12-13)
+
+
+### Bug Fixes
+
+* **config:** fix errors with new app_scripts 0.0.47, and update dep ([813bc70](https://github.com/marcoturi/ionic2-boilerplate/commit/813bc70))
+
+
+
+<a name="0.15.0"></a>
+# [0.15.0](https://github.com/marcoturi/ionic2-boilerplate/compare/v0.14.1...v0.15.0) (2016-11-28)
+
+
+### Bug Fixes
+
+* **yarn:** fix config command in the CI ([6a2ecb4](https://github.com/marcoturi/ionic2-boilerplate/commit/6a2ecb4))
+
+
+### Features
+
+* **ci:** using yarn for CI instead of npm ([6e61c59](https://github.com/marcoturi/ionic2-boilerplate/commit/6e61c59))
+* **dependencies:** add yarnclean ([3911045](https://github.com/marcoturi/ionic2-boilerplate/commit/3911045))
+* **dependencies:** adding yarn support, close [#14](https://github.com/marcoturi/ionic2-boilerplate/issues/14) ([efe2f55](https://github.com/marcoturi/ionic2-boilerplate/commit/efe2f55))
+* **types:** add lodash-es type ([c7df21d](https://github.com/marcoturi/ionic2-boilerplate/commit/c7df21d))
+
+
+
 <a name="0.14.1"></a>
 ## [0.14.1](https://github.com/marcoturi/ionic2-boilerplate/compare/v0.14.0...v0.14.1) (2016-11-18)